[fibonacciDuration]

+++ src/main/java/exercises/tasks/entry/fibonacci/Fibonacci.java
@@ -4,5 +4,5 @@ package exercises.tasks.entry.fibonacci;
  * Created by hofa on 14.02.2018.
  */
 public interface Fibonacci {
-    int count(int value);
+    long count(long value);
 }
