+++ src/main/java/exercises/tasks/entry/fibonacci/Recursive.java
@@ -4,7 +4,7 @@ package exercises.tasks.entry.fibonacci;
  * Created by hofa on 14.02.2018.
  */
 public class Recursive implements Fibonacci {
-    public int count(int value) {
+    public long count(long value) {
         if (value == 1 || value == 2) {
             return 1;
         } else
